requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.50.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
bigcodebench==0.2.5
blake3==1.0.4
bounded-pool-executor==0.0.3
cachetools==5.5.2
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.9.3
cupy-cuda12x==13.4.1
datasets==3.5.0
deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
dnspython==2.7.0
einops==0.8.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
fire==0.7.0
frozenlist==1.6.0
fsspec==2024.12.0
gguf==0.16.3
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.15.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio-client==1.10.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
importlib-metadata==8.0.0
interegular==0.3.3
jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lark==1.2.2
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mistral-common==1.5.4
mistralai==0.4.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multipledispatch==1.0.0
multiprocess==0.70.16
nanobind==2.7.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
openai==1.77.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.6
orjson==3.10.18
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
pip==24.0
pqdm==0.2.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycountry==24.6.1
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pylatexenc==2.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
ray==2.45.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.4
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tempdir==0.7.1
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
tree-sitter==0.24.0
tree-sitter-python==0.23.6
triton==3.2.0
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.5
watchfiles==1.0.5
websockets==15.0.1
wget==3.2
wheel==0.43.0
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.22.0
